requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.122.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
paho-mqtt==2.1.0
postgrest==2.24.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.24.0
sniffio==1.3.1
starlette==0.50.0
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
google-generativeai==0.8.3